按钮样式修改

+++ src/components/admin/post/AdminPostView.vue
@@ -5,7 +5,8 @@
       <template slot="operation" slot-scope="text, record">
         <div class="editable-row-operations">
           <span>
-            <a @click="() => del(record.key)">删除</a>
+<!--            <a @click="() => del(record.key)">删除</a>-->
+            <el-button @click="del(record.key)" type="danger" size="small" icon="el-icon-delete">删除</el-button>
           </span>
         </div>
       </template>
@@ -86,13 +87,26 @@ export default {
       this.$message.success("删除成功", 2);
     },
     del(key) {
-      axios.delete(api.Post + "/" + key, {
-        headers: { Authorization: localStorage.token }
-      }).then(response => {
-        if (response.status == 204) {
-          this.success();
-          this.fetchList();
-        }
+
+      //删除前添加弹窗
+      this.$confirm('是否确定该文章，删除后无法取消?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        axios.delete(api.Post + "/" + key, {
+          headers: { Authorization: localStorage.token }
+        }).then(response => {
+          if (response.status == 204) {
+            this.success();
+            this.fetchList();
+          }
+        });
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        });
       });
     }
   }
